+++ lib/presentation/component/login_component/divider_with_text.dart
